modify sqlite methods
The column order had been rearranged in the json pulling from PLAID. I'm not sure why/how. I rearranged the column order in the sqlite methods section.
This had caused an error because the first import worked but was NULL for transaction ID. So the next time, since all of the new entries were being checked to see "IF NOT IN" list of txn ID's that contained a NULL, none were being selected since their ID's were NULL...
Well it's fixed for now.